% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_readability}
\alias{bootstrap_readability}
\title{compute bootstrapped SEs for readability statistics}
\usage{
bootstrap_readability(x, n = 100, ..., verbose = FALSE)
}
\arguments{
\item{x}{character or \link{corpus} input object for which readability will be computed}

\item{n}{bootstrap replicates}

\item{...}{additional arguments passed to \code{\link{textstat_readability}}}

\item{verbose}{if \code{TRUE} show status messages}
}
\value{
list consisting of three data.frame objects: the computed values on
  the original texts; the computed mean of the replicates; and the standard
  deviations of the replicates
}
\description{
Function to compute bootstrapped mean and SEs for readability statistics.  This is a wrapper around
\code{\link{textstat_readability}}, that redraws the corpus using sentence-level bootstrapping from the
original texts.
}
\examples{
data(data_corpus_inaugural, package = "quanteda")
bootstrap_readability(data_corpus_inaugural[50:58], n = 10,
    measure = c("Flesch", "Flesch.Kincaid"), verbose = TRUE)
}
