% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_input_make.R
\name{cf_input_make}
\alias{cf_input_make}
\title{make dataset for CrowdFlower upload}
\usage{
cf_input_make(..., filename = "CFdata.csv")
}
\arguments{
\item{...}{a series of datasets to be combined into a single dataset for
upload to CF}

\item{filename}{the filename to which the .csv file will be written}
}
\value{
Returns an invisible copy of the data.frame also written to
  \code{filename} in .csv format.
}
\description{
Prepare a dataset for uploading to CrowdFlower, output as a
comma-separated-value (.csv) file.
}
