% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{corpus_sentences_browse}
\alias{corpus_sentences_browse}
\title{inspect the sentences in a corpus}
\usage{
corpus_sentences_browse(x, n = 1000)
}
\arguments{
\item{x}{the corpus to be inspected.}

\item{n}{max number of sentences; default is 1000}
}
\description{
Inspect the sentences in a corpus, in a browser.  Useful for inspecting how sentences are segmented, to see what problems are occurring in the segmentation.
}
\details{
This function segments the corpus by sentences, and displays the results in a browser for inspection.
}
\examples{
\dontrun{
corpus_sentences_browse(data_corpus_pp)
}
}
