% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars_make.R
\name{covars_make}
\alias{covars_make}
\alias{covars_make.snippet}
\alias{covars_make.data.frame}
\alias{covars_make.corpus}
\alias{covars_make.character}
\alias{covars_make_all}
\title{compute text-based variables from text or snippet data}
\usage{
covars_make(x, ...)

\method{covars_make}{snippet}(x, ...)

\method{covars_make}{data.frame}(x, text_field = "text", ...)

\method{covars_make}{corpus}(x, ...)

\method{covars_make}{character}(x, readability_measure = NULL, normalize = TRUE, ...)

covars_make_all(x, ..., dependency = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{snippet data from [snippets_make()] consisting of the
fields `text`, `docID`, and `snippetID`}

\item{...}{arguments passed through to `covars_make_character`}

\item{text_field}{the name of the text field, if a [data.frame], default
is `"text"`}

\item{readability_measure}{additional readability measures passed through in
the `measure` argument passed to
[textstat_readability][quanteda::textstat_readability]. Because our standard input will be
constituent elements rather than indexes, this defaults to `NULL`
indicating that no compound measures will be used.}

\item{normalize}{if `TRUE`, return proportions of words/sentences as
appropriate, instead of raw counts}

\item{dependency}{logical; if `TRUE` parse dependencies}

\item{verbose}{logical; if `TRUE` print status messages}
}
\value{
the data.frame of snippets `x` with added variables.  Note:

`W_wl.Dale.Chall` is the proportion of words *not* in the Dale-Chall word list.
}
\description{
Compute additional variables to snippet data created by
[snippets_make()].  These are based on tokens, types, and various
readability measures.
}
\details{
`covars_make_all` calls `covars_make`,
  [covars_make_baselines()], and [covars_make_pos()],
  returning them as a data.frame.
}
