% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars_make_baselines.R
\name{covars_make_baselines}
\alias{covars_make_baselines}
\alias{covars_make_baselines.snippet}
\alias{covars_make_baselines.data.frame}
\alias{covars_make_baselines.corpus}
\alias{covars_make_baselines.character}
\title{compute mean baseline frequencies from Google ngrams}
\usage{
covars_make_baselines(x, ...)

\method{covars_make_baselines}{snippet}(x, ...)

\method{covars_make_baselines}{data.frame}(x, text_field = "text", ...)

\method{covars_make_baselines}{corpus}(x, ...)

\method{covars_make_baselines}{character}(x, baseline_data = c("brown",
  "google"), baseline_year = 2000, baseline_word = "the", ...)
}
\arguments{
\item{x}{data.frame of results, if already loaded}

\item{...}{additional arguments passed through to other functions}

\item{text_field}{the name of the text field, if a \link{data.frame}, default
is \code{"text"}}

\item{baseline_data}{\code{"brown"}, \code{"google"}, or both (the default) to indicate the
Brown corpus data or Google n-grams data, respectively.}

\item{baseline_year}{the baseline year to choose for reference, a year ending
in 0 from 1790-2000, or \code{NULL} to match a text to its nearest year
(the year information is taken from the \code{textid} that is part of the
Crowdflower data).  Does not apply if \code{baseline_data = "brown"}.}

\item{baseline_word}{the word against which other word frequencies will be
baselined.  This defaults to "the" but can be any word found in the word
frequency tables}
}
\value{
a data.frame suitable for adding to variates for analysis by
  \link[BradleyTerry2]{BTm}
}
\description{
Computes the mean frequencies of terms in a text based on closest match to
the Google unigram corpus, for the decade in which the text was recorded.
}
\examples{
txt <- c(d1 = quanteda::data_char_sampletext,
         d2 = "No if and or but.",
         d3 = "the")
covars_make_baselines(txt)

\dontrun{head(covars_make_baselines(file = "data/CF_results/f921916.csv"))
head(bt_input_make(file = "data/CF_results/f921916.csv",
                   covars = TRUE, readability_measure = "Flesch")$predictors)
}

}
\seealso{
\code{\link{data_matrix_google1grams}},
  \code{\link{data_integer_brownfreq}}
}
