% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars_make_pos.R
\name{covars_make_pos}
\alias{covars_make_pos}
\alias{covars_make_pos.snippet}
\alias{covars_make_pos.corpus}
\alias{covars_make_pos.data.frame}
\alias{covars_make_pos.character}
\title{calculate part-of-speech information from text snippet data}
\usage{
covars_make_pos(x, ...)

\method{covars_make_pos}{snippet}(x, ...)

\method{covars_make_pos}{corpus}(x, ...)

\method{covars_make_pos}{data.frame}(x, text_field = "text", ...)

\method{covars_make_pos}{character}(
  x,
  text_field = "text",
  dependency = TRUE,
  normalize = TRUE,
  ...
)
}
\arguments{
\item{x}{snippet data from [snippets_make()] consisting of the
fields `text`, `docID`, and `snippetID`}

\item{...}{used to pass the `tagset` argument to
[spacyr::spacy_parse()], for example `tagset = "penn"` to specify the Penn Treebank tagset scheme, instead of the Google
universal tagset default.  See [spacyr::spacy_parse()].}

\item{text_field}{the name of the text field, if a [data.frame], default
is `"text"`}

\item{dependency}{logical; if `TRUE` parse dependencies}

\item{normalize}{if `TRUE`, convert pos tag counts to rates}
}
\value{
the data.frame of added variables, consisting of the frequencies of
  parts of speech in each text
}
\description{
Add additional variables consisting of part-of-speech (POS) frequencies to
snippets.
}
\details{
Note that this requires spaCy to be installed (along with Python).
  See the installation instructions at
  <http://github.com/kbenoit/spacyr>.
}
\examples{
\dontrun{
# some examples here
}
}
