% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_BTm_bms}
\alias{data_BTm_bms}
\title{Fitted Bradley-Terry model from Benoit, Munger, and Spirling (2018)}
\format{
Fitted model of type [BTm][BradleyTerry2::BTm]
}
\usage{
data_BTm_bms
}
\description{
Fitted Bradley-Terry model estimates from the "best" model of 
Benoit, Munger, and Spirling (2018).
}
\references{
Benoit, Kenneth, Kevin Munger, and Arthur Spirling. October 30,
  2017.
  "[Measuring
  and Explaining Political Sophistication Through Textual Complexity.](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3062061)"
  London School of Economics and New York University manuscript.
}
\keyword{datasets}
