% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_Crimson}
\alias{data_corpus_Crimson}
\title{Corpus of archives of articles in the Harvard Crimson}
\source{
\url{https://www.thecrimson.com/sitemap/}
}
\description{
A corpus object of articles retrieved from the online archives of the Harvard
Crimson, 1873-2016.  The corpus includes the following
document variable: \describe{\item{date}{Date of the article (formatted as date)} }
}
