% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_fifthgrade}
\alias{data_corpus_fifthgrade}
\title{Corpus of fifth grade texts}
\source{
\url{https://www.ncsu.edu/project/lancet/fifth.htm}
}
\description{
A corpus object of articles taken from a repository of fifth-grade texts
from The Friday Institute for Educational Innovation.
}
