% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_partybroadcasts}
\alias{data_corpus_partybroadcasts}
\title{Corpus of UK political party broadcasts}
\source{
\url{http://www.politicsresources.net/area/uk/peb.htm}
}
\description{
A corpus object UK political party broadcasts, 1961-2004, for the Labour
and Conservative parties.  The corpus includes the following
document variables: \describe{ \item{party}{factor variable, \code{Lab} or \code{Con}}
\item{date}{Date of the broadcast (formatted as date)} }
}
