% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_pp}
\alias{data_corpus_pp}
\title{Corpus of U.S. Presidential Proclamations}
\format{
Includes one docvar, `Year`.
}
\source{
The American Presidency Project, <http://www.presidency.ucsb.edu/sou.php>.
}
\usage{
data_corpus_pp
}
\description{
Corpus of U.S. Presidential Proclamations
}
\note{
To clean the texts for sentence counting, it is recommended to
  pre-process it using the code from the example, e.g.

  `example(data_corpus_pp)`
}
\examples{
\donttest{
data(data_corpus_pp)
patts2remove <- "^Judges|^For the \\\\w+ Ward|By the President|^[\\\\p{Lu}\\\\s]+$"
data_corpus_pp <-
    quanteda::corpus_trimsentences(data_corpus_pp, min_length = 5, 
                                   exclude_pattern = patts2remove)

## inspect the sentences
# corpus_sentences_browse(data_corpus_pp)
}
}
\keyword{datasets}
