% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_presdebates}
\alias{data_corpus_presdebates}
\title{Corpus of segemented US presidential candidate debates}
\source{
The American Presidency Project,
  \url{http://www.presidency.ucsb.edu}.
}
\description{
A corpus object US presidential candidate debates, currently
  just 2016, but we could easily add more. The corpus includes the following
  document variables: \describe{ \item{speaker}{Last name of the speaker}
  \item{party}{Party of the candidate (Democratic or Republican)}
  \item{date}{Date of the debate.} \item{city}{City where the debate took
  place} \item{speakertype}{e.g. \code{"candidate"}, \code{"moderator"},
  etc.} }
}
