% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_presdebates2016}
\alias{data_corpus_presdebates2016}
\title{Corpus of primary debate transcripts from the US Presidential election
  of 2016}
\source{
The American Presidency Project,
  \url{http://www.presidency.ucsb.edu}.
}
\description{
A corpus of primary debate transcripts from the US Presidential
  election of 2016, segmented by speaker, with additional information as docvars.
}
