% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_SCOTUS}
\alias{data_corpus_SCOTUS}
\title{Corpus of Supreme Court decisions 1790-2012}
\source{
\url{https://supreme.justia.com/}, scraped by Tom Clark.
}
\description{
All decisions written by SCOTUS justices, 1790 to 2012
}
\examples{
\dontrun{
# to load this dataset
data_dropbox(data_corpus_SCOTUS)
}
}
