% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dropbox.R
\name{data_dropbox}
\alias{data_dropbox}
\title{load a file from the `data_pkg` Dropbox location}
\usage{
data_dropbox(object, location = paste0(getOption("ROOT_DROPBOX"), "data_pkg/"))
}
\arguments{
\item{object}{the data object to be loaded (not quoted)}

\item{location}{the directory where the large data objects are located.  The
default location is set upon loading the package, by the environment
variable `getOption("ROOT_DROPBOX")`.}
}
\value{
loads the object into the global environment
}
\description{
Loads a data file from the `data_pkg` directory from a user's Dropbox
root folder.  This keeps very large data objects outside of the package, and
off of GitHub.
}
\examples{
\dontrun{
data_dropbox(data_corpus_SCOTUS)
}
}
