% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_integer_brownfreq}
\alias{data_integer_brownfreq}
\title{Matrix of counts of unigrams from Brown corpus}
\description{
Matrix of counts of term frequencies from the Brown corpus, taken from
the \pkg{zipfR} package.
}
