% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_matrix_google1grams}
\alias{data_matrix_google1grams}
\title{Matrix of counts of unigrams from Google 1-gram corpus}
\source{
\url{http://storage.googleapis.com/books/ngrams/books/datasetsv2.html}
}
\description{
Matrix of counts by decade of English language words from Google 1-gram corpus.
}
