% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairSnippets.R
\name{pairs_bridge}
\alias{pairs_bridge}
\title{bridge different pair datasets}
\usage{
pairs_bridge(..., bridge_size = 3)
}
\arguments{
\item{...}{data.frames of pairs created by \code{\link{pairs_regular_make}} or
\code{\link{pairs_gold_make}}}

\item{bridge_size}{how many snippets to pair from one set to another}
}
\value{
a new paired dataset of all combinations of bridging pairs between
  input datasets
}
\description{
Form bridges between datasets by pairing across different sets.  Currently,
this function samples the first snippet from \code{bridge_size} pairs
randomly drawn from each input dataset, and then forms all pairwise
combinations between all of these. This is overkill, perhaps, but ensures
good linkage between the pairs drawn both within and across datasets.
}
\details{
So for instance, with three datasets of 10 pairs each, 3 pairs would be drawn
from each dataset, for 9 total pairs, forming 9 choose 2 =
}
