% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{pairs_gold_browse}
\alias{pairs_gold_browse}
\title{inspect gold snippet pairs}
\usage{
pairs_gold_browse(goldPairs, filename = NULL, ...)
}
\arguments{
\item{goldPairs}{a snippet gold object produced by [pairs_gold_make()]}

\item{filename}{optional filename where the html output will be written; if none
is supplied, writes to an R [tempfile]}

\item{...}{additional arguments passed to [xtable][xtable::xtable]}
}
\description{
Opens gold snippet pairs in a browser for easy inspection.
}
