% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGold.R
\name{pairs_gold_make}
\alias{pairs_gold_make}
\title{create gold snippet pairs for CrowdFlower}
\usage{
pairs_gold_make(x, n.pairs, metric = "Flesch", min.diff.quantile = c(0.1,
  0.9), min.diff.absolute = NULL, screeners = FALSE, verbose = TRUE,
  seed = NULL)
}
\arguments{
\item{x}{snippet pairs}

\item{n.pairs}{how many gold pairs to form}

\item{metric}{measure from \code{\link[quanteda]{textstat_readability}} to use to
compute difference, default is \code{"Flesch"}}

\item{min.diff.quantile}{use this quantile range of difference in the metric
only.  This is a relative measure that defaults to the upper and lower
deciles: \code{c(.1, .9))}}

\item{min.diff.absolute}{form gold pairs with only at least this absolute
value difference in the readability measures of the two texts.  Only one of
\code{min.diff.quantile} and \code{min.diff.absolute} may be specified.}

\item{screeners}{if \code{TRUE}, add screener text with random instructions}

\item{verbose}{if \code{TRUE}, output status messages during processing}

\item{seed}{seed to use for random procedures; set this to ensure
replicability}
}
\value{
a data.frame of gold pairings, with reasons
}
\description{
Create test questions for CrowdFlower from sample of snippet pairs, where the
readability is the most different between pairs.
}
