% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{pairs_regular_browse}
\alias{pairs_regular_browse}
\title{inspect gold snippet pairs}
\usage{
pairs_regular_browse(pairs, filename = NULL, ...)
}
\arguments{
\item{pairs}{snippet pairs created by gold object produced by \code{\link{pairs_regular_make}}}

\item{filename}{optional filename where the html output will be written; if none
is supplied, writes to an R \link{tempfile}}

\item{...}{additional arguments passed to \link[xtable]{xtable}}
}
\description{
Opens gold snippet pairs in a browser for easy inspection.
}
