% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairSnippets.R
\name{pairs_regular_make}
\alias{pairs_regular_make}
\title{form pairs of snippets for comparison}
\usage{
pairs_regular_make(x, n.sample = NULL, n.pairs = nrow(x) - 1, seed = NULL)
}
\arguments{
\item{x}{snippet data.frame}

\item{n.sample}{subset of observations of \code{x} to sample before forming
pairwise comparisons}

\item{n.pairs}{number of pairwise comparisons to form}

\item{seed}{seed to use for random procedures; set this to ensure
replicability}
}
\value{
data.frame of paired snippets, consisting of the elements
  \code{text1}, \code{docID1}, \code{snippetID1}, \code{text2},
  \code{docID2}, and \code{snippetID2}.
}
\description{
Form pairs of snippets for comparison.  By default, form the simple chained
minimum-spannning tree of (\emph{n} - 1), where \emph{n} is the number
snippets, and each snippet is paired with its neighbour.  If \code{n.pairs} >
\emph{n}, then randomly form addition pairs not already in the MST, where
pairs are combinations without respect to order.  Note that the maximum
number of unique pairings is the combination \emph{n} choose 2, or
\emph{n}(\emph{n} - 1)/2.
}
