% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets_make.R
\name{snippets_make}
\alias{snippets_make}
\alias{snippets_make.character}
\alias{snippets_make.corpus}
\title{create text segments for comparisons}
\usage{
snippets_make(x, ...)

\method{snippets_make}{character}(x, ...)

\method{snippets_make}{corpus}(x, nsentences = 1, minchar = 100,
  maxchar = 350, ...)
}
\arguments{
\item{x}{input objects, texts or a quanteda corpus}

\item{...}{not used}

\item{nsentences}{length in sentences of each snippet}

\item{minchar}{minimum length in characters of a snippet}

\item{maxchar}{maximum length in characters of a snippet}
}
\value{
a data.frame of snippets, with a unique indentifier
}
\description{
Create text segments of a defined length, for comparisons.
}
