% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophistication.R
\docType{package}
\name{sophistication-package}
\alias{sophistication}
\alias{sophistication-package}
\title{Functions to help measure textual sophistication}
\description{
Provides functionality to calibrate and estimate measures of textual
sophistication and reading difficulty. Used in Benoit, Munger and Spirling
(2016).
}
\section{Source code and additional information}{


\url{http://github.com/kbenoit/sophistication}
}

