% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars_make.R
\name{covars_make}
\alias{covars_make}
\alias{covars_make.snippet}
\alias{covars_make.data.frame}
\alias{covars_make.corpus}
\alias{covars_make.character}
\alias{covars_make_all}
\title{compute text-based variables from text or snippet data}
\usage{
covars_make(x, ...)

\method{covars_make}{snippet}(x, ...)

\method{covars_make}{data.frame}(x, text_field = "text", ...)

\method{covars_make}{corpus}(x, ...)

\method{covars_make}{character}(x, readability_measure = NULL, normalize = TRUE, ...)

covars_make_all(x, ..., dependency = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{snippet data from \code{\link[=snippets_make]{snippets_make()}} consisting of the
fields \code{text}, \code{docID}, and \code{snippetID}}

\item{...}{arguments passed through to \code{covars_make_character}}

\item{text_field}{the name of the text field, if a \link{data.frame}, default
is \code{"text"}}

\item{readability_measure}{additional readability measures passed through in
the \code{measure} argument passed to
\link[quanteda.textstats:textstat_readability]{quanteda.textstats::textstat_readability}. Because our standard input will be
constituent elements rather than indexes, this defaults to \code{NULL}
indicating that no compound measures will be used.}

\item{normalize}{if \code{TRUE}, return proportions of words/sentences as
appropriate, instead of raw counts}

\item{dependency}{logical; if \code{TRUE} parse dependencies}

\item{verbose}{logical; if \code{TRUE} print status messages}
}
\value{
the data.frame of snippets \code{x} with added variables.  Note:

\code{W_wl.Dale.Chall} is the proportion of words \emph{not} in the Dale-Chall word list.
}
\description{
Compute additional variables to snippet data created by
\code{\link[=snippets_make]{snippets_make()}}.  These are based on tokens, types, and various
readability measures.
}
\details{
\code{covars_make_all} calls \code{covars_make},
\code{\link[=covars_make_baselines]{covars_make_baselines()}}, and \code{\link[=covars_make_pos]{covars_make_pos()}},
returning them as a data.frame.
}
