% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars_make_pos.R
\name{covars_make_pos}
\alias{covars_make_pos}
\alias{covars_make_pos.snippet}
\alias{covars_make_pos.corpus}
\alias{covars_make_pos.data.frame}
\alias{covars_make_pos.character}
\title{calculate part-of-speech information from text snippet data}
\usage{
covars_make_pos(x, ...)

\method{covars_make_pos}{snippet}(x, ...)

\method{covars_make_pos}{corpus}(x, ...)

\method{covars_make_pos}{data.frame}(x, text_field = "text", ...)

\method{covars_make_pos}{character}(x, text_field = "text",
  dependency = TRUE, normalize = TRUE, ...)
}
\arguments{
\item{x}{snippet data from \code{\link{snippets_make}} consisting of the
fields \code{text}, \code{docID}, and \code{snippetID}}

\item{...}{used to pass the \code{tagset} argument to
\code{\link[spacyr]{spacy_parse}}, for example \code{tagset = "penn"} to specify the Penn Treebank tagset scheme, instead of the Google
universal tagset default.  See \code{\link[spacyr]{spacy_parse}}.}

\item{text_field}{the name of the text field, if a \link{data.frame}, default
is \code{"text"}}

\item{dependency}{logical; if \code{TRUE} parse dependencies}

\item{normalize}{if \code{TRUE}, convert pos tag counts to rates}
}
\value{
the data.frame of added variables, consisting of the frequencies of
  parts of speech in each text
}
\description{
Add additional variables consisting of part-of-speech (POS) frequencies to
snippets.
}
\details{
Note that this requires spaCy to be installed (along with Python).
  See the installation instructions at
  \url{http://github.com/kbenoit/spacyr}.
}
\examples{
\dontrun{
# some examples here
}
}
