% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_eo}
\alias{data_corpus_eo}
\title{Corpus of Executive orders}
\format{includes integer \code{Year}}
\source{
\url{http://www.presidency.ucsb.edu/executive_orders.php}
}
\usage{
data_corpus_eo
}
\description{
Executive orders from 1826 to 2017
}
\keyword{datasets}
