% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_presdebates2016}
\alias{data_corpus_presdebates2016}
\title{Corpus of primary debate transcripts from the US Presidential election
  of 2016}
\format{The corpus includes the following
  document variables: \describe{ \item{speaker}{Last name of the speaker}
  \item{party}{Party of the candidate (Democratic or Republican)}
  \item{date}{Date of the debate.} \item{city}{City where the debate took
  place} \item{speakertype}{e.g. \code{"candidate"}, \code{"moderator"},
  etc.} }}
\source{
The American Presidency Project,
  \url{http://www.presidency.ucsb.edu}.
}
\description{
A corpus of primary debate transcripts from the US Presidential
  election of 2016, segmented by speaker, with additional information as docvars.
}
