% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_make.R
\name{pairs_regular_make}
\alias{pairs_regular_make}
\title{form pairs of snippets for comparison}
\usage{
pairs_regular_make(x, n.sample = NULL, n.pairs = nrow(x) - 1, seed = NULL)
}
\arguments{
\item{x}{snippet data.frame}

\item{n.sample}{subset of observations of `x` to sample before forming
pairwise comparisons}

\item{n.pairs}{number of pairwise comparisons to form}

\item{seed}{seed to use for random procedures; set this to ensure
replicability}
}
\value{
data.frame of paired snippets, consisting of the elements
  `text1`, `docID1`, `snippetID1`, `text2`,
  `docID2`, and `snippetID2`.
}
\description{
Form pairs of snippets for comparison.  By default, form the simple chained
minimum-spannning tree of (*n* - 1), where *n* is the number
snippets, and each snippet is paired with its neighbour.  If `n.pairs` >
*n*, then randomly form addition pairs not already in the MST, where
pairs are combinations without respect to order.  Note that the maximum
number of unique pairings is the combination *n* choose 2, or
*n*(*n* - 1)/2.
}
