% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophistication-package.R
\docType{package}
\name{sophistication-package}
\alias{sophistication}
\alias{sophistication-package}
\title{Functions to help measure textual sophistication}
\description{
Provides functionality to calibrate and estimate measures of textual
sophistication and reading difficulty. Used in Benoit, Munger and Spirling
(2019).
}
\section{Source code and additional information}{


<http://github.com/kbenoit/sophistication>
}

\references{
Benoit, Kenneth, Kevin Munger, and Arthur Spirling. (2019). "Measuring and
 Explaining Political Sophistication Through Textual Complexity." American
 Journal of Political Science 63(2, April): 491–508. doi: 10.1111/ajps.12423
}
